Update fixtures to support Fuel-10.0
- Updated scripts to generate fixtures for 10.0
- Added multirack support and fixtures
- Uploaded new fixtures 10.0 build #260

+++ utils/generate_yamls.sh
@@ -5,8 +5,42 @@
 # and unused (not assigned to any env) nodes.
 #
 
+CWD=$(cd `dirname $0` && pwd -P)
+
 mkdir ./yamls
 rm -f ./yamls/*
+
+function generate_fake_nodes_fixtures {
+  # $1 - first IP of admin network to start generate nodes from
+  # $2 - number of nodes to generate
+  # $3 - name of fixture
+  $CWD/generate_nodes_fixtures.rb $1 $2 > $CWD/fixtures/${3}.json
+}
+
+function create_fake_nodes {
+  manage.py loaddata $CWD/fixtures/${1}.json
+}
+
+function clean_fake_nodes {
+  fuel nodes | grep -q 'discover | fnode-' &&
+    fuel nodes | awk '/discover \| fnode-/{ print $1 }' | xargs fuel node --delete-from-db --force --node
+}
+
+function admin_net_tpl {
+  fuel network-group list | awk '/^1 /{print $9}' | sed -e 's/\.[[:digit:]]\+$//'
+}
+
+function id_of_role {
+  env=$1
+  role=$2
+  yaml=`grep -rl primary-controller deployment_$env/*yaml | head -n1`
+  ruby -ryaml -rpp -e '
+  astute = YAML.load(File.read(ARGV[0]))
+  role = ARGV[1]
+  node = astute["network_metadata"]["nodes"].find{|key, hash| hash["node_roles"].include?("#{role}") }
+  puts node.last["uid"]
+  ' $yaml $role
+}
 
 function enable_ceph {
   fuel env --attributes --env $1 --download
@@ -88,7 +122,12 @@ function enable_vms_conf {
 }
 
 function list_free_nodes {
-  fuel nodes 2>/dev/null | grep discover | grep None | awk '{print $1}'
+  # list unused nodes from the list of fake nodes
+  if [ -n "$1" ] ; then
+    fuel nodes 2>/dev/null | grep discover | grep None | grep 'fnode-' | grep $1 | awk '{print $1}'
+  else
+    fuel nodes 2>/dev/null | grep discover | grep None | grep 'fnode-' | awk '{print $1}'
+  fi
 }
 
 function save_yamls {
@@ -102,7 +141,8 @@ function envid {
 
 function store_yamls {
   for role in $3 ; do
-    src=`ls deployment_$1/${role}_*.yaml | head -n1`
+    id=`id_of_role $1 $role`
+    src="deployment_$1/${id}.yaml"
     cp $src ./yamls/$2-$role.yaml
   done
 }
@@ -111,6 +151,10 @@ function generate_yamls {
   env=`envid $1`
   name=$2
   roles=($3)
+
+  # Create fake nodes for our fixtures
+  generate_fake_nodes_fixtures $admin_first_ip 10 default_nodegroup
+  create_fake_nodes default_nodegroup
 
   if [ "${name/ceph}" != "$name" ] ; then
     enable_ceph $env
@@ -134,8 +178,18 @@ function generate_yamls {
     enable_public_ssl $env
   fi
 
+  if [ "${name/multirack}" != "$name" ] ; then
+    # move controllers to custom node group
+    for id in `list_free_nodes 9.9.9` ; do
+      if [ "${roles[0]}" = "controller" ] ; then
+        fuel --env $env node set --node $id --role ${roles[0]}
+        roles=("${roles[@]:1}")
+      fi
+    done
+  fi
+
   for id in `list_free_nodes` ; do
-    if ! [ -z "${roles[0]}" ] ; then
+    if [ -n "${roles[0]}" ] ; then
       fuel --env $env node set --node $id --role ${roles[0]}
       roles=("${roles[@]:1}")
       sleep 1
@@ -153,11 +207,28 @@ function generate_yamls {
 
 function clean_env {
   env=`envid $1`
-  fuel env --delete --env $env
-  rm -rf "cluster_$1"
-  rm -rf "deployment_$env"
-  sleep 80
+  if fuel env --env $env | grep $1 ; then
+    fuel env --delete --env $env
+    rm -rf "cluster_$env"
+    rm -rf "deployment_$env"
+    sleep 80
+  fi
+  clean_fake_nodes
 }
+
+function add_nodegroup {
+  env=`envid $1`
+  name=$2
+
+  fuel --env $env nodegroup --create --name $name
+}
+
+clean_fake_nodes
+sleep 1
+
+# Get some context
+admin_net=$(admin_net_tpl)
+admin_first_ip="${admin_net}.100"
 
 # Neutron vlan ceph
 fuel env --create --name test_neutron_vlan --rel 2 --net vlan
@@ -179,10 +250,10 @@ fuel env --create --name test_neutron_tun --rel 2 --net tun
 generate_yamls 'test_neutron_tun' 'neut_tun.ceph.murano.sahara.ceil' 'controller controller compute ceph-osd ceph-osd mongo mongo' 'primary-controller controller compute ceph-osd primary-mongo mongo'
 clean_env 'test_neutron_tun'
 
-# Neutron tun ironic
-fuel env --create --name test_neutron_tun --rel 2 --net tun
-generate_yamls 'test_neutron_tun' 'neut_tun.ironic' 'controller ironic' 'primary-controller ironic'
-clean_env 'test_neutron_tun'
+# Neutron vlan ironic
+fuel env --create --name test_neutron_vlan --rel 2 --net vlan
+generate_yamls 'test_neutron_vlan' 'neut_tun.ironic' 'controller ironic' 'primary-controller ironic'
+clean_env 'test_neutron_vlan'
 
 # Neutron-l3ha tun + nova_quota
 fuel env --create --name test_neutron_tun --rel 2 --net tun
@@ -194,7 +265,11 @@ fuel env --create --name test_neutron_tun --rel 2 --net tun
 generate_yamls 'test_neutron_tun' 'neut_tun.vms_conf' 'virt compute' 'virt'
 clean_env 'test_neutron_tun'
 
-# Neutron tun, addons, ceph, public and hotizon ssl
+# Multirack, Neutron tun, addons, ceph, public and horizon ssl
 fuel env --create --name test_neutron_tun --rel 2 --net tun
-generate_yamls 'test_neutron_tun' 'neut_tun.murano.sahara.ceil.public_ssl' 'controller controller mongo mongo compute ceph-osd ceph-osd' 'primary-controller compute ceph-osd primary-mongo'
+add_nodegroup 'test_neutron_tun' 'custom_group1'
+generate_fake_nodes_fixtures 9.9.9.150 5 custom_nodegroup
+create_fake_nodes custom_nodegroup
+generate_yamls 'test_neutron_tun' 'neut_tun.multirack.murano.sahara.ceil.ceph.public_ssl' 'controller controller controller mongo mongo compute ceph-osd ceph-osd' 'primary-controller compute ceph-osd primary-mongo'
 clean_env 'test_neutron_tun'
+
